refactor: reorganize project structure by moving source files into psutils directory
