+++ psutils/mem.go
@@ -1,4 +1,5 @@
-package psutilgolangmohamedfadel
+package psutils
+
 
 import (
 	"strconv"
